+++ crate/cli/src/lib.rs
@@ -1,0 +1,5 @@
+//! Command line interface for the peace automation framework.
+//!
+//! This is enabled though the `"cli"` feature on the `peace` crate.
+
+pub mod output;