+++ src/lib.rs
@@ -5,6 +5,10 @@
 pub use miette;
 
 pub use peace_cfg as cfg;
+#[cfg(feature = "cli")]
+pub use peace_cli as cli;
+#[cfg(feature = "cli")]
+pub use peace_cli_model as cli_model;
 pub use peace_cmd as cmd;
 pub use peace_cmd_model as cmd_model;
 pub use peace_cmd_rt as cmd_rt;