+++ workspace_tests/src/cli/output.rs
@@ -4,8 +4,6 @@ mod cli_md_presenter;
 mod cli_output;
 mod cli_output_builder;
 mod cli_output_target;
-mod output_format;
-mod output_format_parse_error;
 
 cfg_if::cfg_if! {
     if #[cfg(feature = "output_progress")] {