+++ workspace_tests/src/lib.rs
@@ -17,6 +17,7 @@ pub(crate) mod mock_item;
 
 // `peace` test modules
 mod cfg;
+mod cli;
 mod cmd;
 mod cmd_model;
 mod cmd_rt;